Output DTSTART correctly for RFC
- I was not outputting the RRULE correctly when RFC compatibility was
set for DTSTART. According to the spec there should be a semi-colon iff
there is a TZID attribute on the DTSTART key. Iff there is no TZID param
after the DTSTART key, there should be a colon, followed by the value of
the DTSTART key.

eg.

 DTSTART;TZID=America/New_York:20180101T090000 - correct
 DTSTART:20180101T090000Z - correct
 DTSTART:TZID=America/New_York:20180101T090000 - not correct

- This was causing issues passing the RRULE strings around between
different RRULE libraries (PHP/JS versions of RRULE libraries)