Limiting BYSECOND to 59
This is narrower than the RFC, which specifies 60 to be used for leap seconds,
but we apparently loop forever if we allow this input (truly horrifying for
servers). Also restored a test that I removed because I thought it was testing
some error condition. It seems the purpose of it was actually just to run to
exhaustion and it was purposefully using invalid input so there'd be no matches.